+++ src/Buckle/Compiler/CodeAnalysis/CodeGeneration/CodeGenerator.CallKind.cs
@@ -1,0 +1,10 @@
+
+namespace Buckle.CodeAnalysis.CodeGeneration;
+
+internal sealed partial class CodeGenerator {
+    private enum CallKind : byte {
+        Call,
+        CallVirt,
+        ConstrainedCallVirt,
+    }
+}
