+++ src/Buckle/Compiler/CodeAnalysis/CodeGeneration/CodeGenerator.UseKind.cs
@@ -1,0 +1,10 @@
+
+namespace Buckle.CodeAnalysis.CodeGeneration;
+
+internal sealed partial class CodeGenerator {
+    private enum UseKind : byte {
+        Unused,
+        UsedAsValue,
+        UsedAsAddress,
+    }
+}
