+++ src/Buckle/Compiler/CodeAnalysis/CodeGeneration/LabelInfo.cs
@@ -1,0 +1,4 @@
+
+namespace Buckle.CodeAnalysis.CodeGeneration;
+
+internal abstract class LabelInfo { }
