+++ src/Buckle/Compiler/CodeAnalysis/CodeGeneration/VariableDefinition.cs
@@ -1,0 +1,4 @@
+
+namespace Buckle.CodeAnalysis.CodeGeneration;
+
+internal abstract class VariableDefinition { }
