+++ src/Buckle/Compiler/CodeAnalysis/Compilation/ModuleBuilder.cs
@@ -1,0 +1,8 @@
+
+namespace Buckle.CodeAnalysis;
+
+internal abstract class ModuleBuilder {
+    internal bool hasGeneratedGlobalsClass;
+
+    internal abstract void EmitGlobalsClass();
+}
