+++ src/Buckle/Compiler/CodeAnalysis/Emitting/CecilLabelInfo.cs
@@ -1,0 +1,11 @@
+using Buckle.CodeAnalysis.CodeGeneration;
+
+namespace Buckle.CodeAnalysis.Emitting;
+
+internal sealed class CecilLabelInfo : LabelInfo {
+    internal CecilLabelInfo(int targetInstructionIndex) {
+        this.targetInstructionIndex = targetInstructionIndex;
+    }
+
+    internal int targetInstructionIndex { get; }
+}
