--- src/Buckle/Compiler/CodeAnalysis/Emitting/ILEmitter.CodeGenerator.CallKind.cs
@@ -1,11 +1,0 @@
-namespace Buckle.CodeAnalysis.Emitting;
-
-internal sealed partial class ILEmitter {
-    internal sealed partial class CodeGenerator {
-        private enum CallKind : byte {
-            Call,
-            CallVirt,
-            ConstrainedCallVirt,
-        }
-    }
-}
