--- src/Buckle/Compiler/CodeAnalysis/Emitting/ILEmitter.CodeGenerator.UseKind.cs
@@ -1,11 +1,0 @@
-namespace Buckle.CodeAnalysis.Emitting;
-
-internal sealed partial class ILEmitter {
-    internal sealed partial class CodeGenerator {
-        private enum UseKind : byte {
-            Unused,
-            UsedAsValue,
-            UsedAsAddress,
-        }
-    }
-}
