--- src/Buckle/Compiler/CodeAnalysis/Evaluating/Executor.cs
@@ -1,15 +1,0 @@
-using Buckle.CodeAnalysis.Binding;
-
-namespace Buckle.CodeAnalysis.Evaluating;
-
-/// <summary>
-/// Executes the program by converting it into an executable and running that executable.
-/// </summary>
-internal static class Executor {
-    /// <summary>
-    /// Executes the given <see cref="BoundProgram" />.
-    /// </summary>
-    internal static void Execute(BoundProgram program, string[] arguments) {
-        // TODO
-    }
-}
