+++ src/Buckle/Compiler/CodeAnalysis/Evaluating/Interpreter.cs
@@ -1,5 +1,3 @@
-using System.Collections.Generic;
-using Buckle.CodeAnalysis.Symbols;
 using Buckle.CodeAnalysis.Syntax;
 using Buckle.CodeAnalysis.Text;
 using Shared;
